Move resource setup to parent Octavia controller
- Moved SSH key management to octavia controller
- Moved Certs management to octavia controller
- Names of certs secret and ca passphrase secret are now fix and
depend on the name of the octavia instance (which is "octavia" normally)
