+++ api/v1beta1/amphoracontroller_types.go
@@ -71,26 +71,6 @@ type OctaviaAmphoraControllerSpec struct {
 	Secret string `json:"secret"`
 
 	// +kubebuilder:validation:Optional
-	// LoadBalancerCerts - Secret containing certs for securing communication with amphora based Load Balancers
-	LoadBalancerCerts string `json:"certsSecret"`
-
-	// +kubebuilder:validation:Optional
-	// +kubebuilder:default=octavia-ca-passphrase
-	// Name of secret containing passphrase for the CA private keys
-	CAKeyPassphraseSecret string `json:"certsPassphraseSecret"`
-
-	// +kubebuilder:validation:Optional
-	// LoadBalancerSSHPubKey - The name of the ConfigMap containing the
-	// pubilc key for connecting to the amphorae via SSH
-	LoadBalancerSSHPubKey string `json:"sshPubkey,omitempty"`
-
-	// +kubebuilder:validation:Optional
-	// LoadBalancerSSHPrivKey - The name of the secret that will be used to
-	// store the private key for connecting to amphorae via SSH. Only used
-	// when no public key was defined and a new key pair needs to be generated.
-	LoadBalancerSSHPrivKey string `json:"sshPrivkeySecret,omitempty"`
-
-	// +kubebuilder:validation:Optional
 	// +kubebuilder:default={database: OctaviaDatabasePassword, service: OctaviaPassword}
 	// PasswordSelectors - Selectors to identify the DB and AdminUser password from the Secret
 	PasswordSelectors PasswordSelector `json:"passwordSelectors,omitempty"`
