+++ config/crd/bases/octavia.openstack.org_octaviaamphoracontrollers.yaml
@@ -74,15 +74,6 @@ spec:
                   - name
                   type: object
                 type: array
-              certsPassphraseSecret:
-                default: octavia-ca-passphrase
-                description: Name of secret containing passphrase for the CA private
-                  keys
-                type: string
-              certsSecret:
-                description: LoadBalancerCerts - Secret containing certs for securing
-                  communication with amphora based Load Balancers
-                type: string
               containerImage:
                 description: ContainerImage - Amphora Controller Container Image URL
                 type: string
@@ -227,16 +218,6 @@ spec:
                 default: octavia
                 description: 'ServiceUser - service user name (TODO: beagles, do we
                   need this at all)'
-                type: string
-              sshPrivkeySecret:
-                description: LoadBalancerSSHPrivKey - The name of the secret that
-                  will be used to store the private key for connecting to amphorae
-                  via SSH. Only used when no public key was defined and a new key
-                  pair needs to be generated.
-                type: string
-              sshPubkey:
-                description: LoadBalancerSSHPubKey - The name of the ConfigMap containing
-                  the pubilc key for connecting to the amphorae via SSH
                 type: string
               tenantName:
                 default: service
