+++ config/crd/bases/octavia.openstack.org_octavias.yaml
@@ -68,11 +68,6 @@ spec:
                   - name
                   type: object
                 type: array
-              certsSecret:
-                default: octavia-certs-secret
-                description: LoadBalancerCerts - Secret containing certs for securing
-                  communication with amphora based Load Balancers
-                type: string
               customServiceConfig:
                 default: '# add your customization here'
                 description: CustomServiceConfig - customize the service config using
@@ -483,15 +478,6 @@ spec:
                       - name
                       type: object
                     type: array
-                  certsPassphraseSecret:
-                    default: octavia-ca-passphrase
-                    description: Name of secret containing passphrase for the CA private
-                      keys
-                    type: string
-                  certsSecret:
-                    description: LoadBalancerCerts - Secret containing certs for securing
-                      communication with amphora based Load Balancers
-                    type: string
                   containerImage:
                     description: ContainerImage - Amphora Controller Container Image
                       URL
@@ -638,17 +624,6 @@ spec:
                     default: octavia
                     description: 'ServiceUser - service user name (TODO: beagles,
                       do we need this at all)'
-                    type: string
-                  sshPrivkeySecret:
-                    description: LoadBalancerSSHPrivKey - The name of the secret that
-                      will be used to store the private key for connecting to amphorae
-                      via SSH. Only used when no public key was defined and a new
-                      key pair needs to be generated.
-                    type: string
-                  sshPubkey:
-                    description: LoadBalancerSSHPubKey - The name of the ConfigMap
-                      containing the pubilc key for connecting to the amphorae via
-                      SSH
                     type: string
                   tenantName:
                     default: service
@@ -693,15 +668,6 @@ spec:
                       - name
                       type: object
                     type: array
-                  certsPassphraseSecret:
-                    default: octavia-ca-passphrase
-                    description: Name of secret containing passphrase for the CA private
-                      keys
-                    type: string
-                  certsSecret:
-                    description: LoadBalancerCerts - Secret containing certs for securing
-                      communication with amphora based Load Balancers
-                    type: string
                   containerImage:
                     description: ContainerImage - Amphora Controller Container Image
                       URL
@@ -848,17 +814,6 @@ spec:
                     default: octavia
                     description: 'ServiceUser - service user name (TODO: beagles,
                       do we need this at all)'
-                    type: string
-                  sshPrivkeySecret:
-                    description: LoadBalancerSSHPrivKey - The name of the secret that
-                      will be used to store the private key for connecting to amphorae
-                      via SSH. Only used when no public key was defined and a new
-                      key pair needs to be generated.
-                    type: string
-                  sshPubkey:
-                    description: LoadBalancerSSHPubKey - The name of the ConfigMap
-                      containing the pubilc key for connecting to the amphorae via
-                      SSH
                     type: string
                   tenantName:
                     default: service
@@ -903,15 +858,6 @@ spec:
                       - name
                       type: object
                     type: array
-                  certsPassphraseSecret:
-                    default: octavia-ca-passphrase
-                    description: Name of secret containing passphrase for the CA private
-                      keys
-                    type: string
-                  certsSecret:
-                    description: LoadBalancerCerts - Secret containing certs for securing
-                      communication with amphora based Load Balancers
-                    type: string
                   containerImage:
                     description: ContainerImage - Amphora Controller Container Image
                       URL
@@ -1058,17 +1004,6 @@ spec:
                     default: octavia
                     description: 'ServiceUser - service user name (TODO: beagles,
                       do we need this at all)'
-                    type: string
-                  sshPrivkeySecret:
-                    description: LoadBalancerSSHPrivKey - The name of the secret that
-                      will be used to store the private key for connecting to amphorae
-                      via SSH. Only used when no public key was defined and a new
-                      key pair needs to be generated.
-                    type: string
-                  sshPubkey:
-                    description: LoadBalancerSSHPubKey - The name of the ConfigMap
-                      containing the pubilc key for connecting to the amphorae via
-                      SSH
                     type: string
                   tenantName:
                     default: service
@@ -1133,7 +1068,6 @@ spec:
                   the pubilc key for connecting to the amphorae via SSH
                 type: string
             required:
-            - certsSecret
             - databaseInstance
             - octaviaAPI
             - rabbitMqClusterName
